Enlever ConsoleHelper
Déplacer ses méthodes dans MenuHelper

Enlever le dossier Helper
