+++ TP1Maintenance/Members/Student.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using Helper;
 using Managers;
 
-namespace Members  
+namespace Members
 {
     public class Student : SchoolMember
     {
@@ -69,8 +68,8 @@ namespace Members
         public static Student StudentAttributes()
         {
             SchoolMember member = SchoolMemberManager.BaseMemberAttributes();
-            int grade = ConsoleHelper.AskNumberInput("Enter grade: ");
-            
+            int grade = MenuHelper.AskNumberInput("Enter grade: ");
+
             return new Student(member.Name, member.Address, member.PhoneNumber, grade);
         }
     }
