+++ TP1Maintenance/Program.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Members;
-using Helper;
 using Managers;
 
 public class Program
